implemented level service funcs

+++ AFTAS_Backend/src/main/java/com/youcode/aftas_backend/repositories/LevelRepository.java
@@ -3,5 +3,5 @@ package com.youcode.aftas_backend.repositories;
 import com.youcode.aftas_backend.models.entities.Level;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LevelRepository extends JpaRepository<Level, Long> {
+public interface LevelRepository extends JpaRepository<Level, Integer> {
 }
