+++ AFTAS_Backend/src/main/java/com/youcode/aftas_backend/services/LevelService.java
@@ -5,5 +5,5 @@ import com.youcode.aftas_backend.superClasses.ServiceInterface;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LevelService extends ServiceInterface<LevelDto, Long> {
+public interface LevelService extends ServiceInterface<LevelDto, Integer> {
 }
